+++ src/hub-connection.connection.spec.ts
@@ -37,6 +37,7 @@ describe("HubConnection Specs", () => {
 			});
 
 			afterEach(() => {
+				// SUT.dispose(); // todo: find why this is failing tests
 				conn$$.unsubscribe();
 			});