+++ src/hub-connection.model.ts
@@ -33,7 +33,8 @@ export interface HubConnectionOptions {
 	key: string;
 	endpointUri: string;
 	options?: ConnectionOptions;
-	defaultData?: () => Dictionary<string>; // todo: rename to getDefaultData
+	// todo: rename to getDefaultData or defaultDataFactory () => Dictionary<string> | Promise<Dictionary<string>>
+	defaultData?: () => Dictionary<string>;
 	/** @internal */
 	getData?: () => Dictionary<string>;
 	protocol?: IHubProtocol;